# backend/docs/swagger.yaml
basePath: /
definitions:
  controllers.BulkDeleteRequest:
    properties:
      ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - ids
    type: object
  controllers.CreateURLRequest:
    properties:
      expires_at:
        type: string
      url:
        type: string
    required:
    - url
    type: object
...
  controllers.Request:
    properties:
      custom_code:
        type: string
      expiry:
        type: string
      long_url:
        type: string
    required:
    - long_url
    type: object
  controllers.UpdateURLRequest:
    properties:
      long_url:
        type: string
    required:
    - long_url
    type: object
...
  models.CoreURLResponse:
    properties:
      short_code:
        type: string
      short_url:
        type: string
    type: object
...
  models.HTTPError:
    properties:
      error:
        example: invalid credentials
        type: string
    type: object
...
  models.MessageSuccess:
    properties:
      message:
        example: operation successful
        type: string
    type: object
  models.PaginatedURLResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.URL'
        type: array
      meta:
        $ref: '#/definitions/models.PaginationMeta'
    type: object
...
  models.PaginationMeta:
    properties:
      limit:
        example: 50
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 150
        type: integer
      total_pages:
        example: 3
        type: integer
    type: object
...
  models.URL:
    properties:
      clicks:
        type: integer
      created_at:
        type: string
      expiry:
        type: string
      id:
        description: go for uuid later
        type: integer
      long_url:
        type: string
      short_url:
        type: string
      user_id:
        type: integer
    required:
    - long_url
    - short_url
    type: object
  models.URLShortenSuccess:
    properties:
      success:
        example: url inserted to db
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: URL shortener with Redis caching, sliding-window limits, buffered analytics
    and Prometheus metrics.
  title: URL Shortener
  version: "1.0"
paths:
...
  /api/urls:
    post:
      consumes:
      - application/json
      description: Creates an anonymous URL; optional expires_at must be in the future.
      parameters:
      - description: URL to shorten
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateURLRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CoreURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/models.HTTPError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/models.HTTPError'
      summary: Create a short URL
      tags:
      - urls
...
  /api/v1/shorten:
    post:
      consumes:
      - application/json
      parameters:
      - description: URL and optional premium custom code; expiry is currently ignored
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.URLShortenSuccess'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.HTTPError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.HTTPError'
      security:
      - Bearer: []
      summary: Create an authenticated short URL
      tags:
      - legacy URLs
  /api/v1/urls:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PaginatedURLResponse'
      security:
      - Bearer: []
      summary: List the current user's short URLs
      tags:
      - legacy URLs
  /api/v1/urls/{id}:
    delete:
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageSuccess'
      security:
      - Bearer: []
      summary: Delete an owned URL and invalidate its cache
      tags:
      - legacy URLs
    put:
      consumes:
      - application/json
      parameters:
      - description: URL ID
        in: path
        name: id
        required: true
        type: integer
      - description: New destination
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageSuccess'
      security:
      - Bearer: []
      summary: Update an owned URL and invalidate its cache
      tags:
      - legacy URLs
  /api/v1/urls/bulk:
    delete:
      consumes:
      - application/json
      parameters:
      - description: URL IDs
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.BulkDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.MessageSuccess'
      security:
      - Bearer: []
      summary: Delete owned URLs in a batch
      tags:
      - legacy URLs
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"